requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
decorator==5.1.1
distro==1.9.0
emoji==2.10.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==0.8.3
faiss-cpu==1.7.4
filetype==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
ipython==8.20.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.1
langchain-community==0.0.16
langchain-core==0.1.17
langchain-openai==0.0.5
langdetect==1.0.9
langsmith==0.0.85
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.3
openai==1.10.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
pip==23.3.1
prompt-toolkit==3.0.43
protobuf==4.25.2
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.15.1
pypdf==4.0.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-iso639==2024.1.2
python-magic==0.4.27
pytz==2023.4
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
selenium==4.16.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.2.0
streamlit==1.29.0
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.7.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
tzlocal==5.2
unstructured==0.12.2
unstructured-client==0.16.0
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.5
wheel==0.41.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
